changed return type of CountAllMilitiaInFiveSectors() from UINT8 to UINT16 to fix an overflow during militia reinforcement
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@9400 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Strategic/Reinforcement.cpp
@@ -207,9 +207,9 @@ UINT8 GetAdjacentSectors( UINT8 pSectors[4], INT16 sSectorX, INT16 sSectorY )
 	return ubCounter;
 }
 
-UINT8 CountAllMilitiaInFiveSectors(INT16 sMapX, INT16 sMapY)
+UINT16 CountAllMilitiaInFiveSectors(INT16 sMapX, INT16 sMapY)
 {
-	UINT8 ubResult = 0;
+	UINT16 ubResult = 0;
 	UINT16 pusMoveDir[4][3];
 	UINT8 ubDirNumber, ubIndex;
